pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [
    {name = "Kartik",email = "kartik.pawade.work@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.133.1,<0.134.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "supabase (>=2.28.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
